+++ processor/src/test/resources/expected/com/example/arez/React4j_AutorunArezReactComponent.java
@@ -31,7 +31,8 @@ abstract class React4j_AutorunArezReactComponent extends AutorunArezReactCompone
   }
 
   @Override
-  protected void reportPropsChanged(@Nullable final JsPropertyMap<Object> nextProps) {
+  protected boolean shouldComponentUpdate(@Nullable final JsPropertyMap<Object> nextProps) {
+    return false;
   }
 
   @Override