+++ processor/src/test/resources/expected/com/example/arez/React4j_ComponentWithCallback.java
@@ -39,7 +39,8 @@ abstract class React4j_ComponentWithCallback extends ComponentWithCallback {
   }
 
   @Override
-  protected void reportPropsChanged(@Nullable final JsPropertyMap<Object> nextProps) {
+  protected boolean shouldComponentUpdate(@Nullable final JsPropertyMap<Object> nextProps) {
+    return false;
   }
 
   @Nonnull