+++ processor/src/test/resources/expected/com/example/arez/React4j_OverridingComponentDidUpdateComponent.java
@@ -30,7 +30,8 @@ abstract class React4j_OverridingComponentDidUpdateComponent extends OverridingC
   }
 
   @Override
-  protected void reportPropsChanged(@Nullable final JsPropertyMap<Object> nextProps) {
+  protected boolean shouldComponentUpdate(@Nullable final JsPropertyMap<Object> nextProps) {
+    return false;
   }
 
   @JsType(